+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Upgrade java base image to 11.0.5-4
 - If the CES admin group changes, the old admin group loses all permissions in Jenkins now
 
+Attention! DO NOT change the CES global admin group and upgrade the dogu at the same time!
+If you have changed the global admin group, restart the Jenkins dogu before upgrading it!
+
 ### Fixed
 - The init090NORMmavenautoinstall.groovy script does no longer crash when the default `M3` maven installer has been adjusted by the Jenkins administrator; #70